[FLINK-14845][runtime][table] Move block compression utils from table module to runtime module.
