+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/compression/BlockCompressor.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.runtime.compression;
+package org.apache.flink.runtime.io.compression;
 
 import java.nio.ByteBuffer;
