+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/compression/Lz4BlockCompressionFactory.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.runtime.compression;
+package org.apache.flink.runtime.io.compression;
 
 /**
  * Implementation of {@link BlockCompressionFactory} for Lz4 codec.
@@ -28,7 +28,7 @@ public class Lz4BlockCompressionFactory implements BlockCompressionFactory {
 	 * We put two integers before each compressed block, the first integer represents the compressed
 	 * length of the block, and the second one represents the original length of the block.
 	 */
-	static final int HEADER_LENGTH = 8;
+	public static final int HEADER_LENGTH = 8;
 
 	@Override
 	public BlockCompressor getCompressor() {
