+++ flink-runtime/src/main/resources/META-INF/NOTICE
@@ -10,6 +10,7 @@ This project bundles the following dependencies under the Apache Software Licens
 - io.netty:netty:3.10.6.Final
 - org.apache.zookeeper:zookeeper:3.4.10
 - org.uncommons.maths:uncommons-maths:1.2.2a
+- org.lz4:lz4-java:1.5.0
 
 This project bundles io.netty:netty:3.10.6.Final from which it inherits the following notices:
